Adjust style for new plugin directory
WordPress.org updated it's plugin directory design. This makes a few updates for styles so that the button is still just as wonderful as it once ones.

Now uses the "button" and "button-large" classes used on the page  on the link so the style should be a bit more persistant.